pyproject.toml
[tool.poetry]
name = "geo-service"
version = "0.1.0"
description = ""
authors = ["JKearnsl <boyarshinovn@gmail.com>", "georgiisafonkin <georgiisafonk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"

# Utils
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
SQLAlchemy = "^2.0.25"
aiofiles = "^23.2.1"
obspy = "^1.4.0"
pandas = "^2.2.0"
APScheduler = "^3.10.4"
httpx = "^0.26.0"
aiomultiprocess = "^0.9.0"
h5py = "^3.10.0"
pyvista = "^0.43.2"
trame = "^3.5.2"
trame-vtk = "^2.8.0"
pyproject-toml = "^0.0.10"
python-jose = "^3.3.0"
passlib = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "src/geo"
